Refactor object detection to use multiprocess instead of thread

+++ examples/lab1part2_object_detector.py
@@ -20,6 +20,7 @@ import time
 import numpy as np
 import cv2
 import atexit
+import os
 from tflite_support.task import core
 from tflite_support.task import processor
 from tflite_support.task import vision
@@ -50,47 +51,60 @@ class Object_detector:
     self.target_fps = 3.5
     self.start_time = time.time()
 
-    # Create capture object
-    self.cap = cv2.VideoCapture(self.camera_id)
-    self.cap.set(cv2.CAP_PROP_FRAME_WIDTH, width)
-    self.cap.set(cv2.CAP_PROP_FRAME_HEIGHT, height)
-
     # Stream visualization parameters
     self.row_size = 20  # pixels
     self.left_margin = 24  # pixels
     self.text_color = (0, 0, 255)  # red
-    self.font_size = 2
-    self.font_thickness = 2
+    self.font_size = 1
+    self.font_thickness = 1
     self.fps_avg_frame_count = 10
     self.window_name = 'object_detector'
 
     # Bounding box visualization parameters
     self._MARGIN = 10  # pixels
     self._ROW_SIZE = 10  # pixels
-    self._FONT_SIZE = 2
-    self._FONT_THICKNESS = 2
+    self._FONT_SIZE = 1
+    self._FONT_THICKNESS = 1
     self._TEXT_COLOR = (0, 0, 255)  # red
+
+    # Flag to stop run_blocking
+    self.should_stop = False
+
+    # Only one process can have the capture open due to the physical constraint 
+    # of having one camera.
+    self.cap = None
+    self.cap_owner_pid = None
+
+    # Register the cleanup method to make sure
+    # that we release the capture when object exits
+    atexit.register(self.cleanup)
+
+  def start_capture(self):
+    # This should only be called once
+    assert(self.cap_owner_pid == None)
+
+    # This process becomes the owner of the capture object
+    self.cap_owner_pid = os.getpid()
+    print(f"Starting capture on process {self.cap_owner_pid}")
+
+    # Create capture object
+    self.cap = cv2.VideoCapture(self.camera_id)
+    self.cap.set(cv2.CAP_PROP_FRAME_WIDTH, self.width)
+    self.cap.set(cv2.CAP_PROP_FRAME_HEIGHT, self.height)
 
     # Initialize the object detection model
     base_options = core.BaseOptions(
-        file_name=model, use_coral=enable_edgetpu, num_threads=num_threads)
+        file_name=self.model, use_coral=self.enable_edgetpu, num_threads=self.num_threads)
     detection_options = processor.DetectionOptions(
         max_results=3, score_threshold=0.3)
     options = vision.ObjectDetectorOptions(
         base_options=base_options, detection_options=detection_options)
     self.detector = vision.ObjectDetector.create_from_options(options)
 
-    # Flag to stop run_periodic_blocking
-    self.should_stop = False
-
-    # Register the cleanup method to make sure
-    # that we release the capture when object exits
-    atexit.register(self.cleanup)
-
-  def run_periodic_blocking(self, period, detected_objects, threshold):
+  def run_blocking(self, detected_objects, threshold):
+    self.start_capture()
     while self.should_stop == False:
       self.process_frame(detected_objects, threshold)
-      time.sleep(period)
 
   def process_frame(self, detected_objects, threshold=0.75):
     """
@@ -132,7 +146,8 @@ class Object_detector:
                 self.font_size, self.text_color, self.font_thickness)
 
     # Todo: make detected_objects thread-safe
-    detected_objects.clear()
+    detected_objects[:] = [] 
+
     # Build up the return list of detections
     for detection in detection_result.detections:
       category = detection.categories[0]
@@ -185,7 +200,8 @@ class Object_detector:
     self.start_time = time.time()
 
   def cleanup(self):
-    print("Safely closing camera capture...")
-    self.cap.release()
-    cv2.destroyAllWindows()
-    print("Done")
+    if os.getpid() == self.cap_owner_pid:
+      print("Safely closing camera capture...")
+      self.cap.release()
+      cv2.destroyAllWindows()
+      print("Done")
